chore: Move snakecase.go to stringutil subfolder.
