+++ tm2/pkg/amino/genproto/genproto.go
@@ -16,6 +16,7 @@ import (
 	"time"
 
 	"github.com/gnolang/gno/tm2/pkg/amino"
+	"github.com/gnolang/gno/tm2/pkg/amino/genproto/stringutil"
 	"github.com/gnolang/gno/tm2/pkg/amino/pkg"
 )
 
@@ -206,7 +207,7 @@ func (p3c *P3Context) GenerateProto3MessagePartial(p3doc *P3Doc, rt reflect.Type
 		p3Field := P3Field{
 			Repeated: fp3IsRepeated,
 			Type:     fp3,
-			Name:     ToLowerSnakeCase(field.Name),
+			Name:     stringutil.ToLowerSnakeCase(field.Name),
 			Number:   field.FieldOptions.BinFieldNum,
 		}
 		p3msg.Fields = append(p3msg.Fields, p3Field)
